+++ CHANGELOG.md
@@ -63,6 +63,7 @@ Inventory causes an error. Refs UIIN-2012.
 * Fix filters list doesn't update when user back to "Browse contributors" pane. Refs UIIN-2069.
 * Applied filters don't reset when user select "Browse contributors" search option. Refs UIIN-2068.
 * Create Jest/RTL test for PrimaryToggleButton. Refs UIIN-1666.
+* Adjust query parameters on the subject browse form. Refs UIIN-2075.
 
 ## [9.0.0](https://github.com/folio-org/ui-inventory/tree/v9.0.0) (2022-03-03)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v8.0.0...v9.0.0)